Move parseXML into own package
