+++ ecfr-parser/parseXML/parseXML.go
@@ -1,0 +1,202 @@
+package parseXML
+
+import (
+	"encoding/xml"
+	"fmt"
+	"io"
+	"log"
+)
+
+type Part struct {
+	XMLName   xml.Name     `xml:"DIV5"`
+	Name      string       `xml:"N,attr"`
+	Type      string       `xml:"TYPE,attr"`
+	Header    string       `xml:"HEAD"`
+	Authority string       `xml:"AUTH>PSPACE"`
+	Source    string       `xml:"SOURCE>PSPACE"`
+	Children  PartChildren `xml:",any"`
+}
+
+type Subpart struct {
+	Header   string          `xml:"HEAD"`
+	Type     string          `xml:"TYPE,attr"`
+	Children SubpartChildren `xml:",any"`
+}
+
+type SubjectGroup struct {
+	Header string `xml:"HEAD"`
+	Type   string          `xml:"TYPE,attr"`
+	Children SubjectGroupChildren `xml:",any"`
+}
+
+type Section struct {
+	Type string `xml:"TYPE,attr"`
+	Header   string          `xml:"HEAD"`
+	Children SectionChildren `xml:",any"`
+}
+
+type Paragraph struct {
+	Type string
+	Content string `xml:",innerxml"`
+}
+
+type Extract struct {
+	Type string
+	Content string `xml:",innerxml"`
+}
+
+type Citation struct {
+	Type string
+	Content string `xml:",innerxml"`
+}
+
+type Source struct {
+	Type string
+	Header string `xml:"HED"`
+	Content string `xml:"PSPACE"`
+}
+
+type SectionAuthority struct {
+	Type string
+	Content string `xml:",innerxml"`
+}
+
+type FlushParagraph struct {
+	Type string
+	Content string `xml:",innerxml"`
+}
+
+type Image struct {
+	Type string
+	Source string `xml:"src,attr"`
+}
+
+type PartChildren []interface{}
+type SubpartChildren []interface{}
+type SectionChildren []interface{}
+type SubjectGroupChildren []interface{}
+
+func (c *PartChildren) UnmarshalXML(d *xml.Decoder, start xml.StartElement) error {
+	switch start.Name.Local {
+	case "DIV6":
+		child := &Subpart{}
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	case "DIV8":
+		child := &Section{}
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	default:
+		//return fmt.Errorf("Unknown XML type in Part: %+v", start)
+		fmt.Printf("Unknown XML type in Part: %+v\n", start)
+		d.Skip()
+	}
+
+	return nil
+}
+
+func (c *SubpartChildren) UnmarshalXML(d *xml.Decoder, start xml.StartElement) error {
+	switch start.Name.Local {
+	case "DIV8":
+		child := &Section{}
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	case "DIV7":
+		child := &SubjectGroup{}
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	case "SOURCE":
+		child := &Source{Type: "Source"}
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	default:
+		//return fmt.Errorf("Unknown XML type in Subpart: %+v", start)
+		fmt.Printf("Unknown XML type in Subpart: %+v\n", start)
+		d.Skip()
+	}
+	return nil
+}
+
+func (c *SubjectGroupChildren) UnmarshalXML(d *xml.Decoder, start xml.StartElement) error {
+	switch start.Name.Local {
+	case "DIV8":
+		child := &Section{}
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	default:
+		//return fmt.Errorf("Unknown XML type in Subpart: %+v", start)
+		fmt.Printf("Unknown XML type in Subject Group: %+v\n", start)
+		d.Skip()
+	}
+	return nil
+}
+
+func (c *SectionChildren) UnmarshalXML(d *xml.Decoder, start xml.StartElement) error {
+	switch start.Name.Local {
+	case "P":
+		child := &Paragraph{ Type: "Paragraph"}
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	case "FP":
+		child := &FlushParagraph{ Type: "FlushParagraph"}
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	case "img":
+		child := &Image{ Type: "Image"}
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	case "EXTRACT":
+		child := &Extract{ Type: "Extract" }
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	case "CITA":
+		child := &Citation{ Type: "Citation" }
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	case "SECAUTH":
+		child := &SectionAuthority{ Type: "SectionAuthority" }
+		if err := d.DecodeElement(child, &start); err != nil {
+			return err
+		}
+		*c = append(*c, child)
+	default:
+		//return fmt.Errorf("Unknown XML type in Section: %+v", start)
+		fmt.Printf("Unknown XML type in Section: %+v\n", start)
+		d.Skip()
+	}
+
+	return nil
+}
+
+func ParsePart(b io.ReadCloser) (*Part, error) {
+	d := xml.NewDecoder(b)
+
+	p := &Part{}
+	log.Println("[DEBUG] Decoding part")
+	if err := d.Decode(p); err != nil {
+		return p, err
+	}
+	return p, nil
+}
